// frontend/src/client/types.gen.ts
export type ActivityCreate = {
    type: ActivityType;
    subject: string;
    description?: (string | null);
    due_date?: (string | null);
    completed?: boolean;
    company_id: string;
    contact_id?: (string | null);
    deal_id?: (string | null);
};

export type ActivityPublic = {
    type: ActivityType;
    subject: string;
    description?: (string | null);
    due_date?: (string | null);
    completed?: boolean;
    id: string;
    company_id: string;
    contact_id: (string | null);
    deal_id: (string | null);
    owner_id: string;
    tenant_id: string;
    created_at?: (string | null);
};

export type ActivityType = 'call' | 'email' | 'meeting' | 'task';